pyproject.toml
[project]
name = "danchoicloud"
version = "0.1.0"
description = "Telegram bot for dan choi cloud"
readme = "README.md"
authors = [
    { name = "hoangndst", email = "hoangndst@gmail.com" },
]
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "python-telegram-bot>=20.0",
    "httpx>=0.25.0",
    "APScheduler>=3.10.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "tabulate>=0.9.0",
    "python-dateutil>=2.8.0",
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
